requirements.txt
absl-py==2.3.1
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
ecdsa==0.19.1
fastapi==0.116.1
flatbuffers==25.2.10
fonttools==4.59.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
h11==0.16.0
h5py==3.14.0
idna==3.10
jax==0.7.0
jaxlib==0.7.0
keras==3.11.1
kiwisolver==1.4.8
kombu==5.5.4
libclang==18.1.1
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.3
namex==0.1.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
passlib==1.7.4
pillow==11.3.0
prompt_toolkit==3.0.51
protobuf==4.25.8
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
redis==6.4.0
requests==2.32.4
rich==14.1.0
rsa==4.9.1
scipy==1.16.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
SQLAlchemy==2.0.42
starlette==0.47.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
